Add files via upload
Build a Web-side with Web 3.0 with blockchain and
next.js , solidity sanity i.o. Sample twitter web-site used for social 
networking platform where people can express their thoughts in short 
posts which people can read and share! Read a News some information. 
This project consept is using for future coding technology it call as web 
3.0. Today, we are going to make the web-site for tweet consept using 
Next JS in the frontend along with Html, Css , javascript. And back-end 
with sanity i.o. Tailwind Css is designsystem implementation in pure Css. 
It is also configurable. It Gives Developers Super power. It allows them to 
Build websites with a clean consistent UI. MetaMask is the trailblazing 
tool enabling user interactions and experience on Web3. It is currently 
available as a browser extension and as a mobile app on both Android and 
iOS devices. The purpose of this documentation is to illustrate how to build 
a dapp with MetaMask. Sanity i.o use Back-end Database. Next.js gives
you the best developer experience with all the features you need for
production: hybrid static & server rendering, TypeScript support, smart 
bundling. This Web-site very style and look cool color.